CRUD Examen Hematologico finalizado
-Correcciones en heces microscopico

-Correcciones en heces quimico

-Falta correcciones de anexo y heces quimico que se pertenezcan los ids

+++ src/Controller/ExamenHematologicoController.php
@@ -24,9 +24,49 @@ class ExamenHematologicoController extends AbstractController
      */
     public function index(ExamenHematologicoRepository $examenHematologicoRepository,ExamenSolicitado $examen_solicitado, Security $AuthUser): Response
     {
-        return $this->render('examen_hematologico/index.html.twig', [
-            'examen_hematologicos' => $examenHematologicoRepository->findAll(),
-        ]);
+        if($AuthUser->getUser()->getRol()->getNombreRol() != 'ROLE_SA'){
+            if($AuthUser->getUser()->getClinica()->getId() == $examen_solicitado->getCita()->getExpediente()->getUsuario()->getClinica()->getId()){
+                if($examen_solicitado->getCita()->getExpediente()->getHabilitado()){
+
+                    $em = $this->getDoctrine()->getManager();
+                    $RAW_QUERY = "SELECT examen.* FROM `examen_hematologico` as examen WHERE examen_solicitado_id =".$examen_solicitado->getId().";";
+                    $statement = $em->getConnection()->prepare($RAW_QUERY);
+                    $statement->execute();
+                    $result = $statement->fetchAll();
+
+                    return $this->render('examen_hematologico/index.html.twig', [
+                        'examen_hematologicos'          => $result,
+                        'user'                          => $AuthUser,
+                        'examen_solicitado'             => $examen_solicitado,
+                    ]);
+
+                }else{
+                    $this->addFlash('fail','Este paciente no esta habilitado, para poder hacer uso de el consulte con su superior para habilitar el paciente');
+                    return $this->redirectToRoute('home');
+                }
+            }else{
+                $this->addFlash('fail','Error, este registro puede que no exista o no le pertenece');
+                return $this->redirectToRoute('home');
+            }  
+        }
+
+        if($examen_solicitado->getCita()->getExpediente()->getHabilitado()){
+            $em = $this->getDoctrine()->getManager();
+            $RAW_QUERY = "SELECT examen.* FROM `examen_hematologico` as examen WHERE examen_solicitado_id =".$examen_solicitado->getId().";";
+            $statement = $em->getConnection()->prepare($RAW_QUERY);
+            $statement->execute();
+            $result = $statement->fetchAll();
+
+            return $this->render('examen_hematologico/index.html.twig', [
+                'examen_hematologicos'    => $result,
+                'user'                          => $AuthUser,
+                'examen_solicitado'             => $examen_solicitado,
+            ]);
+        }else{
+            $this->addFlash('fail','Este paciente no esta habilitado, para poder hacer uso de el consulte con su superior para habilitar el paciente');
+            return $this->redirectToRoute('home');
+        }
+
     }
 
     /**
@@ -34,22 +74,130 @@ class ExamenHematologicoController extends AbstractController
      */
     public function new(Request $request,ExamenSolicitado $examen_solicitado, Security $AuthUser): Response
     {
-        $examenHematologico = new ExamenHematologico();
-        $form = $this->createForm(ExamenHematologicoType::class, $examenHematologico);
-        $form->handleRequest($request);
+        if($AuthUser->getUser()->getRol()->getNombreRol() != 'ROLE_SA'){
+            if($AuthUser->getUser()->getClinica()->getId() == $examen_solicitado->getCita()->getExpediente()->getUsuario()->getClinica()->getId()){
+                if($examen_solicitado->getCita()->getExpediente()->getHabilitado()){
+                    $editar = false;
+                    $examenHematologico = new ExamenHematologico();
+                    $form = $this->createForm(ExamenHematologicoType::class, $examenHematologico);
+                    $form->handleRequest($request);
+                    if ($form->isSubmitted() && $form->isValid()) {
+                        if($form["tipoSerie"]->getData() != ""){
+                            if($form["unidad"]->getData() != ""){
+                                if($form["valorReferencia"]->getData() != ""){
+                                    //PROCESAMIENTO DE DATOS
+                                    $entityManager = $this->getDoctrine()->getManager();
+                                    $examenHematologico->setExamenSolicitado($examen_solicitado);
+                                    $entityManager->persist($examenHematologico);
+                                    $entityManager->flush();
+                                    //FIN DE PROCESAMIENTO DE DATOS
+                                }else{
+                                    $this->addFlash('fail', 'Error, la serie no puede ir vacio');
+                                    return $this->render('examen_hematologico/new.html.twig', [
+                                        'examen_hematologico' => $examenHematologico,
+                                        'examen_solicitado' => $examen_solicitado,
+                                        'editar'            => $editar,
+                                        'form' => $form->createView(),
+                                    ]);
+                                }
+                            }else{
+                                $this->addFlash('fail', 'Error, la unidad no puede ir vacio');
+                                return $this->render('examen_hematologico/new.html.twig', [
+                                    'examen_hematologico' => $examenHecesQuimico,
+                                    'examen_solicitado' => $examen_solicitado,
+                                    'editar'            => $editar,
+                                    'form' => $form->createView(),
+                                ]);
+                            }
+                        }else{
+                            $this->addFlash('fail', 'Error, el valor de referencia no puede ir vacio, si no hay resultados que asignar por favor asigne " * "');
+                            return $this->render('examen_hematologico/new.html.twig', [
+                                'examen_hematologico' => $examenHecesQuimico,
+                                'examen_solicitado' => $examen_solicitado,
+                                'editar'            => $editar,
+                                'form' => $form->createView(),
+                            ]);
+                        }
+                        $this->addFlash('success', 'Examen añadido con exito');
+                        return $this->redirectToRoute('examen_hematologico_index',['examen_solicitado' => $examen_solicitado->getId()]);
+                    }
+                    
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->persist($examenHematologico);
-            $entityManager->flush();
-
-            return $this->redirectToRoute('examen_hematologico_index');
+                    return $this->render('examen_hematologico/new.html.twig', [
+                        'examen_hematologico' => $examenHematologico,
+                        'examen_solicitado' => $examen_solicitado,
+                        'editar'            => $editar,
+                        'form' => $form->createView(),
+                    ]);
+                }else{
+                    $this->addFlash('fail','Este paciente no esta habilitado, para poder hacer uso de el consulte con su superior para habilitar el paciente');
+                    return $this->redirectToRoute('home');
+                }
+            }else{
+                $this->addFlash('fail','Error, este registro puede que no exista o no le pertenece');
+                return $this->redirectToRoute('home');
+            }  
         }
 
-        return $this->render('examen_hematologico/new.html.twig', [
-            'examen_hematologico' => $examenHematologico,
-            'form' => $form->createView(),
-        ]);
+
+
+        if($examen_solicitado->getCita()->getExpediente()->getHabilitado()){
+            $editar = false;
+            $examenHematologico = new ExamenHematologico();
+            $form = $this->createForm(ExamenHematologicoType::class, $examenHematologico);
+            $form->handleRequest($request);
+            if ($form->isSubmitted() && $form->isValid()) {
+                if($form["tipoSerie"]->getData() != ""){
+                    if($form["unidad"]->getData() != ""){
+                        if($form["valorReferencia"]->getData() != ""){
+                            //PROCESAMIENTO DE DATOS
+                            $entityManager = $this->getDoctrine()->getManager();
+                            $examenHematologico->setExamenSolicitado($examen_solicitado);
+                            $entityManager->persist($examenHematologico);
+                            $entityManager->flush();
+                            //FIN DE PROCESAMIENTO DE DATOS
+                        }else{
+                            $this->addFlash('fail', 'Error, la serie no puede ir vacio');
+                            return $this->render('examen_hematologico/new.html.twig', [
+                                'examen_hematologico' => $examenHematologico,
+                                'examen_solicitado' => $examen_solicitado,
+                                'editar'            => $editar,
+                                'form' => $form->createView(),
+                            ]);
+                        }
+                    }else{
+                        $this->addFlash('fail', 'Error, la unidad no puede ir vacio');
+                        return $this->render('examen_hematologico/new.html.twig', [
+                            'examen_hematologico' => $examenHecesQuimico,
+                            'examen_solicitado' => $examen_solicitado,
+                            'editar'            => $editar,
+                            'form' => $form->createView(),
+                        ]);
+                    }
+                }else{
+                    $this->addFlash('fail', 'Error, el valor de referencia no puede ir vacio, si no hay resultados que asignar por favor asigne " * "');
+                    return $this->render('examen_hematologico/new.html.twig', [
+                        'examen_hematologico' => $examenHecesQuimico,
+                        'examen_solicitado' => $examen_solicitado,
+                        'editar'            => $editar,
+                        'form' => $form->createView(),
+                    ]);
+                }
+                $this->addFlash('success', 'Examen añadido con exito');
+                return $this->redirectToRoute('examen_hematologico_index',['examen_solicitado' => $examen_solicitado->getId()]);
+            }
+            
+
+            return $this->render('examen_hematologico/new.html.twig', [
+                'examen_hematologico' => $examenHematologico,
+                'examen_solicitado' => $examen_solicitado,
+                'editar'            => $editar,
+                'form' => $form->createView(),
+            ]);
+        }else{
+            $this->addFlash('fail','Este paciente no esta habilitado, para poder hacer uso de el consulte con su superior para habilitar el paciente');
+            return $this->redirectToRoute('home');
+        }
     }
 
     /**
@@ -57,31 +205,89 @@ class ExamenHematologicoController extends AbstractController
      */
     public function show(ExamenHematologico $examenHematologico,ExamenSolicitado $examen_solicitado, Security $AuthUser): Response
     {
+        if($AuthUser->getUser()->getRol()->getNombreRol() != 'ROLE_SA'){
+            if($AuthUser->getUser()->getClinica()->getId() == $examen_solicitado->getCita()->getExpediente()->getUsuario()->getClinica()->getId() && $examenHematologico->getExamenSolicitado()->getId() == $examen_solicitado->getId()){
+                if($examen_solicitado->getCita()->getExpediente()->getHabilitado()){
+                    return $this->render('examen_hematologico/show.html.twig', [
+                        'examen_hematologico' => $examenHematologico,
+                        'examen_solicitado' => $examen_solicitado,
+                    ]);
+                }else{
+                    $this->addFlash('fail','Este paciente no esta habilitado, para poder hacer uso de el consulte con su superior para habilitar el paciente');
+                    return $this->redirectToRoute('home');
+                }
+            }else{
+                $this->addFlash('fail','Error, este registro puede que no exista o no le pertenece');
+                return $this->redirectToRoute('home');
+            }  
+        }
         return $this->render('examen_hematologico/show.html.twig', [
             'examen_hematologico' => $examenHematologico,
+            'examen_solicitado' => $examen_solicitado,
         ]);
     }
 
     /**
      * @Route("/{id}/{examen_solicitado}/edit", name="examen_hematologico_edit", methods={"GET","POST"})
      */
+
     public function edit(Request $request, ExamenHematologico $examenHematologico,ExamenSolicitado $examen_solicitado, Security $AuthUser): Response
-    {
-        $form = $this->createForm(ExamenHematologicoType::class, $examenHematologico);
-        $form->handleRequest($request);
+    {   
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $this->getDoctrine()->getManager()->flush();
+        if($AuthUser->getUser()->getRol()->getNombreRol() != 'ROLE_SA'){
+            if($AuthUser->getUser()->getClinica()->getId() == $examen_solicitado->getCita()->getExpediente()->getUsuario()->getClinica()->getId() && $examenHematologico->getExamenSolicitado()->getId() == $examen_solicitado->getId()){
+                if($examen_solicitado->getCita()->getExpediente()->getHabilitado()){
+                    $editar = true;
+                    $form = $this->createForm(ExamenHematologicoType::class, $examenHematologico);
+                    $form->handleRequest($request);
 
-            return $this->redirectToRoute('examen_hematologico_index', [
-                'id' => $examenHematologico->getId(),
-            ]);
+                    if ($form->isSubmitted() && $form->isValid()) {
+                        $this->getDoctrine()->getManager()->flush();
+                        $this->addFlash('success', 'Examen modificado con exito');
+                        return $this->redirectToRoute('examen_hematologico_index', [
+                            'id' => $examenHematologico->getId(),
+                            'examen_solicitado' => $examen_solicitado->getId(),
+                        ]);
+                    }
+                    return $this->render('examen_hematologico/edit.html.twig', [
+                        'examen_hematologico' => $examenHematologico,
+                        'examen_solicitado' => $examen_solicitado,
+                        'editar'            => $editar,
+                        'form' => $form->createView(),
+                    ]);
+                }else{
+                    $this->addFlash('fail','Este paciente no esta habilitado, para poder hacer uso de el consulte con su superior para habilitar el paciente');
+                    return $this->redirectToRoute('home');
+                }
+            }else{
+                $this->addFlash('fail','Error, este registro puede que no exista o no le pertenece');
+                return $this->redirectToRoute('home');
+            }  
         }
+        
+        if($examen_solicitado->getCita()->getExpediente()->getHabilitado()){
+            $editar = true;
+            $form = $this->createForm(ExamenHematologicoType::class, $examenHematologico);
+            $form->handleRequest($request);
 
-        return $this->render('examen_hematologico/edit.html.twig', [
-            'examen_hematologico' => $examenHematologico,
-            'form' => $form->createView(),
-        ]);
+            if ($form->isSubmitted() && $form->isValid()) {
+                $this->getDoctrine()->getManager()->flush();
+                $this->addFlash('success', 'Examen modificado con exito');
+                return $this->redirectToRoute('examen_hematologico_index', [
+                    'id' => $examenHematologico->getId(),
+                    'examen_solicitado' => $examen_solicitado->getId(),
+                ]);
+            }
+            return $this->render('examen_hematologico/edit.html.twig', [
+                'examen_hematologico' => $examenHematologico,
+                'examen_solicitado' => $examen_solicitado,
+                'editar'            => $editar,
+                'form' => $form->createView(),
+            ]);
+        }else{
+            $this->addFlash('fail','Este paciente no esta habilitado, para poder hacer uso de el consulte con su superior para habilitar el paciente');
+            return $this->redirectToRoute('home');
+        }
     }
 
     /**
@@ -89,12 +295,45 @@ class ExamenHematologicoController extends AbstractController
      */
     public function delete(Request $request, ExamenHematologico $examenHematologico,ExamenSolicitado $examen_solicitado, Security $AuthUser): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$examenHematologico->getId(), $request->request->get('_token'))) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->remove($examenHematologico);
-            $entityManager->flush();
+        if($AuthUser->getUser()->getRol()->getNombreRol() != 'ROLE_SA'){
+            if($AuthUser->getUser()->getClinica()->getId() == $examen_solicitado->getCita()->getExpediente()->getUsuario()->getClinica()->getId() && $examenHematologico->getExamenSolicitado()->getId() == $examen_solicitado->getId()){
+                if($examen_solicitado->getCita()->getExpediente()->getHabilitado()){
+                    if ($this->isCsrfTokenValid('delete'.$examenHematologico->getId(), $request->request->get('_token'))) {
+                        $entityManager = $this->getDoctrine()->getManager();
+                        $entityManager->remove($examenHematologico);
+                        $entityManager->flush();
+                    }
+                    $this->addFlash('success', 'Examen eliminado con exito');
+                    return $this->redirectToRoute('examen_hematologico_index',['examen_solicitado' => $examen_solicitado->getId()]);
+                }else{
+                    $this->addFlash('fail','Este paciente no esta habilitado, para poder hacer uso de el consulte con su superior para habilitar el paciente');
+                    return $this->redirectToRoute('home');
+                }
+            }else{
+                $this->addFlash('fail','Error, este registro puede que no exista o no le pertenece');
+                return $this->redirectToRoute('home');
+            }  
         }
 
-        return $this->redirectToRoute('examen_hematologico_index');
+       if($examen_solicitado->getCita()->getExpediente()->getHabilitado()){
+            if ($this->isCsrfTokenValid('delete'.$examenHematologico->getId(), $request->request->get('_token'))) {
+                $entityManager = $this->getDoctrine()->getManager();
+                $entityManager->remove($examenHematologico);
+                $entityManager->flush();
+            }
+            $this->addFlash('success', 'Examen eliminado con exito');
+            return $this->redirectToRoute('examen_hematologico_index',['examen_solicitado' => $examen_solicitado->getId()]);
+        }else{
+            $this->addFlash('fail','Este paciente no esta habilitado, para poder hacer uso de el consulte con su superior para habilitar el paciente');
+            return $this->redirectToRoute('home');
+        }
     }
+
+
+
+
+
+
+
+
 }
